client: collect event_handlers
Collecting event handling related functions, traits in a module
